+++ RELEASES.md
@@ -1299,3 +1299,13 @@ Based on:
 - [typescript v0.6.0-alpha.37] .
 ### Releases
 - [NPM v0.6.0-alpha.37] https://www.npmjs.com/package/cribl-control-plane/v/0.6.0-alpha.37 - .
+
+## 2026-03-10 02:09:16
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.749.1 (2.855.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.0-alpha.38] .
+### Releases
+- [NPM v0.6.0-alpha.38] https://www.npmjs.com/package/cribl-control-plane/v/0.6.0-alpha.38 - .